Tinder For Geese
Geese are sorted by species, determined which pairs are the closest, and
a nest centroid is created from this pair.

This code is not optimized, but it was tested on 1000 geese and exicuted
in under 1 second.

Near future todo:
Write nests to a CSV

Future todo:
Implement Ripley's function in C++
Interfacing with image recognition